+++ lib/src/exception.dart
@@ -1,9 +1,21 @@
 class BlurHashDecodeException implements Exception {
-  BlurHashDecodeException({String message}) : message = message ?? "";
+  BlurHashDecodeException([
+    String? message,
+  ]) : message = message ?? '';
+
   final String message;
+
+  @override
+  String toString() => 'Exception: $message';
 }
 
 class BlurHashEncodeException implements Exception {
-  BlurHashEncodeException({String message}) : message = message ?? "";
+  BlurHashEncodeException([
+    String? message,
+  ]) : message = message ?? '';
+
   final String message;
+
+  @override
+  String toString() => 'Exception: $message';
 }